Drop 'ensure-writable'
This flag is only needed if this module tries to modify files from
previous build steps in-place. As best I can tell, unpacking the Scilab
tarball does not do this. And having it present causes problems:
flatpak-builder fails as follows:

    Cache hit for ncurses, skipping build
    Cache miss, checking out last cache hit
    […]
    FB: Breaking hardlink files/share/terminfo/1
    Error: Can’t recursively copy directory

Actually I think this is a flatpak-builder bug: files/share/terminfo/1
is a directory so can't have any hardlinks to break. But removing this
unneeded flag avoids hitting the bug.